move attestation structs to submodules
